+++ bravado/requests_client.py
@@ -93,6 +93,31 @@ class RequestsClient(HttpClient):
         self.session = requests.Session()
         self.authenticator = None
 
+    @staticmethod
+    def separate_params(request_params):
+        """Splits the passed in dict of request_params into two buckets.
+
+        - sanitized_params are valid kwargs for constructing a
+          requests.Request(..)
+        - misc_options are things like timeouts which can't be communicated
+          to the Requests library via the requests.Request(...) constructor.
+
+        :param request_params: kitchen sink of request params. Treated as a
+            read-only dict.
+        :returns: tuple(sanitized_params, misc_options)
+        """
+        sanitized_params = request_params.copy()
+        misc_options = {}
+
+        if 'connect_timeout' in sanitized_params:
+            misc_options['connect_timeout'] = \
+                sanitized_params.pop('connect_timeout')
+
+        if 'timeout' in sanitized_params:
+            misc_options['timeout'] = sanitized_params.pop('timeout')
+
+        return sanitized_params, misc_options
+
     def request(self, request_params, response_callback=None):
         """
         :param request_params: complete request data.
@@ -101,8 +126,12 @@ class RequestsClient(HttpClient):
         :returns: HTTP Future object
         :rtype: :class: `bravado_core.http_future.HttpFuture`
         """
+        sanitized_params, misc_options = self.separate_params(request_params)
+
         requests_future = RequestsFutureAdapter(
-            self.session, self.authenticated_request(request_params))
+            self.session,
+            self.authenticated_request(sanitized_params),
+            misc_options)
 
         return HttpFuture(
             requests_future,
@@ -168,16 +197,22 @@ class RequestsResponseAdapter(IncomingResponse):
 
 
 class RequestsFutureAdapter(object):
-    """A future which inputs HTTP params"""
+    """Mimics a :class:`concurrent.futures.Future` for the purposes of making
+    HTTP calls with the Requests library in a future-y sort of way.
+    """
 
-    def __init__(self, session, request):
+    def __init__(self, session, request, misc_options):
         """Kicks API call for Requests client
 
         :param session: session object to use for making the request
         :param request: dict containing API request parameters
+        :param misc_options: misc options to apply when sending a HTTP request.
+            e.g. timeout, connect_timeout, etc
+        :type misc_options: dict
         """
         self.session = session
         self.request = request
+        self.misc_options = misc_options
 
     def check_for_exceptions(self, response):
         try:
@@ -185,18 +220,63 @@ class RequestsFutureAdapter(object):
         except Exception as e:
             add_response_detail_to_errors(e)
 
-    def result(self, timeout):
+    def build_timeout(self, result_timeout):
+        """
+        Build the appropriate timeout object to pass to `session.send(...)`
+        based on connect_timeout, the timeout passed to the service call, and
+        the timeout passed to the result call.
+
+        :param result_timeout: timeout that was passed into `future.result(..)`
+        :return: timeout
+        :rtype: float or tuple(connect_timeout, timeout)
+        """
+        # The API provides two ways to pass a timeout :( We're stuck
+        # dealing with it until we're ready to make a non-backwards
+        # compatible change.
+        #
+        #  - If both timeouts are the same, no problem
+        #  - If either has a non-None value, use the non-None value
+        #  - If both have a non-None value, use the greater of the two
+        timeout = None
+        has_service_timeout = 'timeout' in self.misc_options
+        service_timeout = self.misc_options.get('timeout')
+
+        if not has_service_timeout:
+            timeout = result_timeout
+        elif service_timeout == result_timeout:
+            timeout = service_timeout
+        else:
+            if service_timeout is None:
+                timeout = result_timeout
+            elif result_timeout is None:
+                timeout = service_timeout
+            else:
+                timeout = max(service_timeout, result_timeout)
+            log.warn("Two different timeouts have been passed: "
+                     "_request_options['timeout'] = {0} and "
+                     "future.result(timeout={1}). Using timeout of {2}."
+                     .format(service_timeout, result_timeout, timeout))
+
+        # Requests is weird in that if you want to specify a connect_timeout
+        # and idle timeout, then the timeout is passed as a tuple
+        if 'connect_timeout' in self.misc_options:
+            timeout = self.misc_options['connect_timeout'], timeout
+        return timeout
+
+    def result(self, timeout=None):
         """Blocking call to wait for API response
 
-        :param timeout: timeout in seconds to wait for response
-        :type timeout: integer
+        :param timeout: timeout in seconds to wait for response. Defaults to
+            None to wait indefinitely.
+        :type timeout: float
         :return: raw response from the server
         :rtype: dict
         """
         request = self.request
+        prepared_request = self.session.prepare_request(request)
         response = self.session.send(
-            self.session.prepare_request(request),
-            timeout=timeout)
+            prepared_request,
+            timeout=self.build_timeout(timeout))
 
         self.check_for_exceptions(response)