+++ CHANGELOG.rst
@@ -1,3 +1,9 @@
+3.0.0 (2015-XX-XX)
+---------------------
+- Support passing in connect_timeout and timeout via _request_options to the Fido and Requests clients
+- Timeout in HTTPFuture now defaults to None (wait indefinitely) instead of 5s. You should make sure
+  any calls to http_future.result(..) without a timeout are updated accordingly.
+
 2.1.0 (2015-07-20)
 ---------------------
 - Add warning for deprecated operations